backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.3
ultralytics==8.3.50
torch>=2.6.0
torchvision>=0.21.0
pillow==11.0.0
requests==2.32.3
cloudinary==1.41.0
deep-translator==1.11.4
python-multipart==0.0.20
aiofiles==24.1.0
google-generativeai==0.8.5
google-cloud-translate==3.15.0
google-cloud-texttospeech==2.16.0
